fix: move getContractInstanceWithProvider to scripts
