+++ packages/scripts/src/utils/contracts.ts
@@ -1,4 +1,4 @@
-import { ContractName, DeploymentName, getAddress } from "@uma/contracts-node";
+import { ContractName, DeploymentName, getAbi, getAddress } from "@uma/contracts-node";
 const hre = require("hardhat");
 import { Contract } from "ethers";
 import { Provider } from "@ethersproject/abstract-provider";
@@ -25,4 +25,15 @@ export const getContractInstanceByUrl = async <T extends Contract>(
   const networkId = (await provider.getNetwork()).chainId;
   const contractAddress = await getAddress(contractName, Number(networkId));
   return hre.ethers.getContractAt(contractName, contractAddress) as T;
+};
+
+export const getContractInstanceWithProvider = async <T extends Contract>(
+  contractName: ContractName,
+  provider: Provider,
+  address?: string
+): Promise<T> => {
+  const networkId = (await provider.getNetwork()).chainId;
+  const contractAddress = address || (await getAddress(contractName, networkId));
+  const contractAbi = getAbi(contractName);
+  return new Contract(contractAddress, contractAbi, provider) as T;
 };
